add DelegateRoleButton and RevokeRoleButton components; integrate RoleList into RoleSection
