+++ src/features/amUI/common/RoleList/RevokeRoleButton.tsx
@@ -4,36 +4,36 @@ import { Button } from '@altinn/altinn-components';
 
 import type { Party } from '@/rtk/features/lookupApi';
 import { useGetReporteePartyQuery } from '@/rtk/features/lookupApi';
-import { useDelegateMutation } from '@/rtk/features/roleApi';
+import { useRevokeMutation } from '@/rtk/features/roleApi';
 
-import { useSnackbar } from '../../common/Snackbar';
-import { SnackbarDuration, SnackbarMessageVariant } from '../../common/Snackbar/SnackbarProvider';
+import { useSnackbar } from '../Snackbar';
+import { SnackbarDuration, SnackbarMessageVariant } from '../Snackbar/SnackbarProvider';
 
-interface DelegateRoleButtonProps extends ButtonProps {
-  roleId: string;
+interface RevokeRoleButtonProps extends ButtonProps {
+  assignmentId: string;
   roleName: string;
   toParty?: Party;
   fullText?: boolean;
 }
 
-export const DelegateRoleButton = ({
-  roleId,
+export const RevokeRoleButton = ({
+  assignmentId,
   roleName,
   toParty,
   fullText = false,
   disabled,
   variant = 'outline',
   ...props
-}: DelegateRoleButtonProps) => {
+}: RevokeRoleButtonProps) => {
   const { t } = useTranslation();
   const { openSnackbar } = useSnackbar();
   const { data: representingParty } = useGetReporteePartyQuery();
-  const [delegateRole, { isLoading }] = useDelegateMutation();
+  const [revoke, { isLoading }] = useRevokeMutation();
 
   const onClick = () => {
     const snackbar = (isSuccessful: boolean) => {
       const snackbarData = {
-        message: t(isSuccessful ? 'role.role_delegation_success' : 'role.role_delegation_error', {
+        message: t(isSuccessful ? 'role.role_deletion_success' : 'role.role_deletion_error', {
           role: roleName,
           name: toParty?.name,
         }),
@@ -44,9 +44,8 @@ export const DelegateRoleButton = ({
     };
 
     if (representingParty) {
-      delegateRole({
-        to: toParty?.partyUuid || '',
-        roleId: roleId,
+      revoke({
+        assignmentId: assignmentId,
       })
         .unwrap()
         .then(() => {
@@ -64,9 +63,9 @@ export const DelegateRoleButton = ({
       {...props}
       variant={variant}
       onClick={onClick}
-      disabled={isLoading || disabled || !representingParty}
+      disabled={disabled || isLoading || !representingParty}
     >
-      {fullText ? t('common.give_poa') : t('common.give_poa')}
+      {fullText ? t('common.delete_poa') : t('common.delete')}
     </Button>
   );
 };
