+++ src/features/amUI/userRightsPage/RoleSection/roleSection.module.css
@@ -1,17 +1,3 @@
-.roleArea {
-  padding-top: 1.5rem;
-}
-
-.roleList {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-  margin-top: 1.5rem;
-}
-
 .roleSection {
   display: flex;
   flex-direction: column;
